+++ scripts/daemon/stop.js
@@ -6,8 +6,8 @@ var talk = require('./utils/talk');
 module.exports = function (path) {
 	// listener for exceptions
 	gn.on('uncaughtException', function (error) {
-		logger.error(lib.color(path + ' ' + sockName(path), lib.COLORS.RED));
-		gn.exit(error);
+		logger.error(lib.color(path, lib.COLORS.RED));
+		gn.exit();
 	});
 	// stop daemon
 	talk.setup(path, function (isAppRunning) {