+++ scripts/daemon/start.js
@@ -28,11 +28,20 @@ module.exports = function (path, logPath) {
 		}
 		// start daemon
 		var child = run(process.execPath, args, { detached: true, stdio: 'ignore' });
-		console.log(lib.color('Daemon process started', lib.COLORS.GRAY), lib.color(path, lib.COLORS.LIGHT_BLUE));
-		var dies = lib.color(' dies', lib.COLORS.BROWN);
-		var ten = lib.color(' 10 ', lib.COLORS.PURPLE);
-		var exit = lib.color(' exit', lib.COLORS.BROWN);
-		console.log(lib.color('If the application', lib.COLORS.GRAY) + dies + ten + lib.color('times in less than', lib.COLORS.GRAY) + ten + lib.color('seconds, the daemon process will', lib.COLORS.GRAY) + exit);
-		gn.exit();
+		talk.isRunning(function (error, running) {
+			if (error) {
+				return gn.exit(error);
+			}
+			if (running) {
+				console.log(lib.color('Daemon process started', lib.COLORS.GRAY), lib.color(path, lib.COLORS.LIGHT_BLUE));
+				var dies = lib.color(' dies', lib.COLORS.BROWN);
+				var ten = lib.color(' 10 ', lib.COLORS.PURPLE);
+				var exit = lib.color(' exit', lib.COLORS.BROWN);
+				console.log(lib.color('If the application', lib.COLORS.GRAY) + dies + ten + lib.color('times in less than', lib.COLORS.GRAY) + ten + lib.color('seconds, the daemon process will', lib.COLORS.GRAY) + exit);
+			} else {
+				console.error(lib.color('Daemon process failed to start', lib.COLORS.RED), lib.color(path, lib.COLORS.RED));
+			}
+			gn.exit();
+		});
 	});
 };